♻️ Refactor: 이벤트 로직 커스텀 훅으로 분리
